datapoint insertion working

+++ tables.sql
@@ -1,6 +1,6 @@
 CREATE TABLE data_streams (
   id         SERIAL,
-  time_stamp  integer NOT NULL,
+  time_stamp  BIGINT NOT NULL,
   count      integer NOT NULL,
   PRIMARY KEY (id, time_stamp)
 );
@@ -12,5 +12,6 @@ CREATE TABLE graphs (
 
   -- FOREIGN KEY(data_stream_id) REFERENCES data_streams(id)
 
+-- ALTER TABLE data_streams ALTER COLUMN time_stamp TYPE BIGINT;
 -- alter table data_streams drop constraint data_streams_pkey;
 -- alter table data_streams add constraint data_streams_pkey primary key (id, time_stamp);
